Add a recipe/BOM CSV import script, widen assembly amounts to decimal
scripts/import_recipes_csv.py loads a Product SKU/Ingredient Number/Amount
export into assemblies: one assembly per Product SKU, one AssemblyVersion
(named --version-name, "Current" by default) holding a negated line item
per ingredient plus one for the product itself (amount = the sum of that
assembly's ingredient amounts, not renormalized to 100). Items are
insert-ignore - an item that already exists is left untouched rather than
overwritten, since all this export gives for a new one is its SKU.
--created-at lets a historical version be backdated to when that recipe
was actually in effect. Verified against a real local Postgres 16
instance: correct sums, insert-ignore leaves pre-existing item names
alone, and re-running is fully idempotent (skips assemblies/versions that
already exist by name).

Recipe amounts aren't whole numbers (e.g. 26.20 parts per batch), so
assembly_version_items.amount widens from Integer to Numeric(12, 4) -
verified via the same local Postgres instance since this table (0015)
isn't in real use yet.

Also fixes scripts/import_items_csv.py, which still set the old
items.category/subcategory string columns removed by 0014 - it now
resolves (or creates) the matching categories row and sets category_id
instead.

Needs `alembic upgrade head` on deploy (adds 0016).

+++ src/models.py
@@ -270,7 +270,9 @@ class AssemblyVersionItem(Base, TimestampMixin):
         ForeignKey("assembly_versions.assembly_version_id"), nullable=False
     )
     product_id: Mapped[int] = mapped_column(ForeignKey("items.item_id"), nullable=False)
-    amount: Mapped[int] = mapped_column(Integer, nullable=False)
+    # A signed quantity, not necessarily a whole number - recipe amounts
+    # (e.g. 26.20 parts per batch) are fractional.
+    amount: Mapped[Decimal] = mapped_column(Numeric(12, 4), nullable=False)
 
     assembly_version: Mapped["AssemblyVersion"] = relationship(back_populates="items")
     product: Mapped["Item"] = relationship()
